LayoutTests/imported/w3c/web-platform-tests/referrer-policy/gen/srcdoc-inherit.meta/default/script-tag.http-expected.txt

Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/script.py?redirection=keep-origin&action=purge&key=8c137ff6-d40a-4266-93ee-a88fff5dcd3b&path=%2Fmixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/script.py?redirection=no-redirect&action=purge&key=127bf433-9bc2-46c7-8164-a9ae179e1fa1&path=%2Fmixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/script.py?redirection=swap-origin&action=purge&key=fd7e85c3-495a-4ace-aa8b-fde3bbff8dbf&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/script.py?redirection=keep-origin&action=purge&key=1aef30f4-6cb4-4ab7-82ae-9cab5b22dc5d&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/script.py?redirection=no-redirect&action=purge&key=96e34dc3-c688-486f-a26c-b0ef5cda5838&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/script.py?redirection=swap-origin&action=purge&key=42b0e8ef-c2ee-497f-9c63-60855063db47&path=%2Fmixed-content
Blocked access to external URL http://www1.localhost:8800/common/security-features/subresource/script.py?action=purge&key=2df8f323-fa0c-4ef1-ac99-155ebd85fc63&path=%2Fmixed-content
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/script.py?action=purge&key=145e06d3-be89-469e-8aee-fc30d704c8c2&path=%2Fmixed-content

FAIL Referrer Policy: Expects origin for script-tag to cross-http origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: "Error"
FAIL Referrer Policy: Expects origin for script-tag to cross-http origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: "Error"
FAIL Referrer Policy: Expects origin for script-tag to cross-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Error"
FAIL Referrer Policy: Expects origin for script-tag to cross-https origin and keep-origin redirection from http context. promise_test: Unhandled rejection with value: "Error"
FAIL Referrer Policy: Expects origin for script-tag to cross-https origin and no-redirect redirection from http context. promise_test: Unhandled rejection with value: "Error"
FAIL Referrer Policy: Expects origin for script-tag to cross-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Error"
FAIL Referrer Policy: Expects origin for script-tag to same-http origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Error"
FAIL Referrer Policy: Expects origin for script-tag to same-https origin and keep-origin redirection from http context. assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/gen/srcdoc-inherit.meta/default/script-tag.http.html"
FAIL Referrer Policy: Expects origin for script-tag to same-https origin and no-redirect redirection from http context. assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/gen/srcdoc-inherit.meta/default/script-tag.http.html"
FAIL Referrer Policy: Expects origin for script-tag to same-https origin and swap-origin redirection from http context. promise_test: Unhandled rejection with value: "Error"
PASS Referrer Policy: Expects stripped-referrer for script-tag to same-http origin and keep-origin redirection from http context.
PASS Referrer Policy: Expects stripped-referrer for script-tag to same-http origin and no-redirect redirection from http context.
